requirements.txt
Pillow==8.0.1
PyOpenGL==3.1.5
easygui==0.98.1
freetype-py==2.2.0
numpy==1.19.4
pip==20.2.4
pygame==2.0.0
setuptools==50.3.2